+++ mustache-cli/test/errors/sys-errors.t
@@ -2,14 +2,14 @@
   $ echo "" > foo.mustache
 
 Nonexistent json file:
-  $ mustache nonexistent.json foo.mustache
+  $ mustache-ocaml nonexistent.json foo.mustache
   mustache: DATA.json argument: no 'nonexistent.json' file or directory
   Usage: mustache [-I DIR] [--no-working-dir] [OPTION]… DATA.json TEMPLATE.mustache
   Try 'mustache --help' for more information.
   [124]
 
 Nonexistent template file:
-  $ mustache foo.json nonexistent.mustache
+  $ mustache-ocaml foo.json nonexistent.mustache
   mustache: TEMPLATE.mustache argument: no 'nonexistent.mustache' file or
             directory
   Usage: mustache [-I DIR] [--no-working-dir] [OPTION]… DATA.json TEMPLATE.mustache